SC_Unit.h: Replace decltype C cast with static_cast
`decltype( a[0] )` returns a referenced type. This sc_typeof_cast used this to perform a c cast, when converting from a double to a float, producing the wrong value.

This PR creates a new conversion function using static_asserts and static_cast that does the correct behaviour and will break at compile time if the user attempts to pass something unexpected.

An example of a bug caused by this error can be found in the following PR.