ClassLib: Replace `doesNotUnderstand` with `doesNotUnderstandAbout`.
Implement performWithKeys, valueWithKeys and functionPerformWithKeys, along with error checking in Method/FunctionDef.

These three methods are useful when using `doesNotUnderstandAbout` as they provide a way to call a function with both arguments and keyword arguments centralizing all the error checking inside of Method/FunctionDef.

These `*WithKeys` methods take an argumentArray and a keywordArgumentPairs array, then check it against the Method or FunctionDef to ensure the keywords are valid and don't overlap with the non-keyword args. Ultimately, these methods are a way to centralize the error checking.