SCLang and ClassLib: implement ...Kwargs
This commit makes many changes, but because of how the interpreter bootstraps itself, it isn't possible to pull these out into smaller atomic steps that will each pass the CI.

Implement a new syntax for passing keyword arguments.

Update parse node code so it is compiler correctly.

Refactor out putting the arguments onto the stack for a method call so the logic for keyword collisions and other issues is always the same. Further, make this function take an optional environment to look values up in. This is called `prepareArgsForExecute`.

Use this new function to rewrite `executeMethod`. Since the logic for all method calls goes through only one function, `executeMethodWithKeys` can be removed. Do the same for `blockValue`, `blockValueWithKeys`, `blockValueEnvir` and `blockValueEnvirWithKeys`, these other methods can't be removed because they are primitive functions and a stable pointer to them is needed.

Similarly, `doesNotUnderstand` is unified with `doesNotUnderstandWithKeys`, refactored, and made to pass the new keyword argument array correctly.

The logic from `sendMessage`, `sendMessageWithKeys`, `superSendMessage` and `superSendMessageWithKeys` is extracted into one function and deletes the `*WithKeys` functions.

An additional primitive, `objectPerformArgsAndKwArgs` is created that allows you to pass an keyword argument array as this is impossible to do otherwise.

The sc method `performList`'s signature is changed to fix a longstanding bug with keyword arguments and `*[]` array unpacking of arguments.
The cpp function objectperformListWithKeys is also created to fix this.

Now in the class library Object is given a new method, `performWithArgsAndKwArgs` allowing to pass an array to the keyword arguments array.

DoesNotUnderstand is updated to allow keywords to be reported.

Object's `doesNotUnderstand` method is updated to accept keyword arguments and passes these to the error correctly.

TestKwargs.sc is created that tests all this behaviour.

Further, the documentation is updated. This required a new primitive to correctly report how many variable arguments there are.