SCLang: make slot tag fixed width integer
On unix `long` is 64 bits, but on windows `long` is 32.
Throughout the code assumptions are made that assume the size of the `PyrSlot` to be 128bits. This makes that explicit rather than relying on alignment, which I believe was the original intention as otherwise this could easily have been a uint8.