SCLang: add keywords to newCopyArgs as constructors are messy with many instance variables
If you have a class with many instance variables and want to only initialise some of them, you currently need to call setters, which gets messy and can be overloaded.

Now you can call them like so...
```
super.newCopyArgs(1, 2, 3, foo: 10, bar: 20)
```

You can directly expose the constructor like this...
```
*new { |...args, kwargs|
   ^super.performArgs(\newCopyArgs, args, kwargs)
}
```
In future this could become the default implementation of `Object*new`.

This PR applied this to `SynthDef` as an example.