Env:-circle: prepended first level to  levels array.
Previously, this value didn't matter because the erroneous UGen initialization swallowed it and it was never output. Now that that's fixed, it gets output for a single sample on account of the construction of the `first0Then1` initial time mechanism. So better to start at (and repeat) the first level, rather than arbitrarily start at `0.0`, which could be destructive if the SynthDef graph isn't designed around the env having a value of 0 (e.g. a freq arg.)