proj docs: windows build guide: persistent Qt PATH
The environment variables by `set` command only valid during the
`cmd.exe` process. To make the change persistent, there are a few
different approaches. And using `setx` might be the most efficient,
especially, because the doc is already using `cmd.exe`.

Further down in the same doc, it is stressed to put the Qt's executable
path in the PATH environment variable. But it does not mention how.

I moved the command a bit earlier on the doc. Because this particular
command is not required to be executed under the `supercollider/build`
directory.

[skip ci]