SCLang: Too many args passed to certain 'inline' methods (#6302)
When the compile/parser see a method with only a single call, or a call to `super.something`, it 'inlines' it.

For these methods, all the users arguments were being passed through, regardless of how many were declared in the function signature.

This commit checks this, and removes any excess args from the stack.

Closes #1087

authored-by: JordanHendersonMusic