SCLang: add keywords to newCopyArgs as constructors are messy with many instance variables (#6398)
SCLang: add keywords to newCopyArgs as constructors are messy with many instance variables

If you have a class with many instance variables and want to only initialise some of them, you currently need to call setters, which gets messy and can be overloaded.

Now you can call them like so...
```
super.newCopyArgs(1, 2, 3, foo: 10, bar: 20)
```

You can directly expose the constructor like this...
```
*new { |...args, kwargs|
   ^super.performArgs(\newCopyArgs, args, kwargs)
}
```
In future this could become the default implementation of `Object*new`.

This PR applied this to `SynthDef` as an example.


* SCLang: fix non-passed keyword arguments in certain compiler call node optimizations.

It doesn't appear it was possible to trigger this bug before #6339, but it existed.

* Docs: improve text description and add better examples for newCopyArgs
* SCUnitTest: add to newCopyArgs, ensuring bug with optimizations and keywords is avoided.
* Help: reword in Object.schelp
* UnitTest: test kw super.new args with incorrect kwarg

---------

authored-by: JordanHendersonMusic