ClassLib: SynthDef refactor
# This is a **large** change to the SynthDef compiler.

Before this commit there the main issues with the compiler were:

1. The user was required to use the firstArgs operator `<!` to specify external resource ordering. This led to issues where if you read and write from the same buffer, the compiler would change the order.

2. There was no functionally 'correct' implementation for which we could compare the output of the SynthDef compiler against.

3. The logic for sorting was confusing as `width-first` sorting essentially changed the sorting algorithm midway through a sort.

4. The logic of UGen optimisations (e.g., merging `a + b + c` into `Sum3(a, b, c)`) had grown unwieldy and could easily become unpredictable as the order of operations was not defined.

These issues are addressed accordingly.

1. The unaltered user written SynthDef is considered the 'source of truth'. It describes how the SynthDef should 'behave'. Behaviour is defined as how a SynthDef alters external **resources**. A resource is considered some stateful action, for an example, writing or reading to a bus, but also sending OSC messages, and DoneActions between synths. This means any optimisation the SynthDef compiler performs is valid, so long as the behaviour is unaltered. This is reflected in the unit tests where the un-optimised version is null-test compared against the optimised. Unfortunately, there is one exception to this, that is that `PV_UGens` have an optimisation path to insert a `PV_Copy` where needed, this makes the optimiser a necessary part of the SynthDef compiler but is maintained for backward compatibility.

2. Each UGen now declares what resources it touches and how it touches them. Following which, a resource manager is used to create additional edges between UGens. This allows normal topological sorting to be used.

3. For optimisations, UGens are allowed to look at their **direct** inputs and their children (at any level). When a substitution is made, the optimiser returns to the children, and checks again if any optimisations can be performed. This check is performed from the outputs of the SynthDef graph to the inputs, working its way upwards. This ensures the graph is completely optimised and produces consistent results.

## How UGens should be written from now on.

A UGen that doesn't use server resources (e.g., it doesn't write/read to a bus, nor buffer, nor other resource).

```supercollider
SinOsc : UGen {
    resourceManagers { ^[] }
    hasObservableEffect { ^false }
	canBeReplacedByIdenticalCall { ^true }

	...
}
```

And here is an example of a bus reading UGen.
```supercollider
In : UGen {
	resourceManagers { ^[UGenBusResourceManager] }
	busAccessType { ^\read }
	hasObservableEffect { ^false }
	canBeReplacedByIdenticalCall { ^true }
}
```

First the resource manager is defined, the class `UGenBusResourceManager` is instantiated by the SynthDef compiler and when this UGen is created, it tells the manager how it will interact with the bus, via `busAccessType`, and the manager creates edges to other UGens as needed to preserve the correct topological ordering.

`hasObservableEffect` and `canBeReplacedByIdenticalCall` are used in the optimisation path of the SynthDef compiler and allow the UGen to be removed if it isn't connected to any output or de-duplicated where possible. Both of these operations are guaranteed to not effect the 'behaviour' of the synth.

### Resource Mangers

A number of resource managers are provided, along with a method to create your own. Here is their implementation.
```supercollider
UGenMessageResourceManager : UGenResourceManagerSimple {
	*initClass { UGenResourceManager.register(this) }
}

UGenDoneResourceManager : UGenResourceManagerSimple {
	*initClass { UGenResourceManager.register(this) }
}

UGenBufferResourceManager : UGenResourceManagerWithNonCausalModes {
	*initClass { UGenResourceManager.register(this) }
	*new { ^super.new([\read], \bufferAccessType) }
}

UGenBusResourceManager : UGenResourceManagerWithNonCausalModes {
	*initClass { UGenResourceManager.register(this) }
	*new { ^super.new([\read, \write], \busAccessType) }
}

UGenLocalBusResourceManager : UGenResourceManagerWithNonCausalModes {
	*initClass { UGenResourceManager.register(this) }
	*new { ^super.new([\readKr, \writeKr, \readAr, \writeAr], \busAccessType) }
}

UGenRandomResourceManager : UGenResourceManagerWithNonCausalModes {
	*initClass { UGenResourceManager.register(this) }
	*new { ^super.new([\gen], \randomAccessType) }
}

UGenAnalogResourceManager : UGenResourceManagerWithNonCausalModes {
	*initClass { UGenResourceManager.register(this) }
	*new { ^super.new([\read], \analogAccessType) }
}

UGenDiskResourceManager : UGenResourceManagerSimple {
	*initClass { UGenResourceManager.register(this) }
}
```

As can be seen, they fall into two groups, `UGenResourceManagerSimple` and `UGenResourceManagerWithNonCausalModes`.

The former simply creates edges between edge UGen that touches that resource. For example, UGens that have a done action are always required to be evaluated in the order they are written in the SynthDef by the user.

The latter, allows for certain 'modes' of access to be reordered. As an example, if you have a group of `In.ar`s (a `busAccessType` of `\read`) there are no `Out.ar` (a `busAccessType` of `\write`) between them, then it is possible to reorder the `In.ar`s as needed. This becomes useful when trying to limit the number of wirebufs as is necessary on lower memory systems.

Importantly, if a bus UGen provides a `busAccessType` that isn't `\read` nor `\write`, then this is considered the end of the previous group and the ordering is preserved.

### Optimisations

Optimisations fall under three categories, sorting, common-sub-expression-elimination (de-duplication), and substitutions.

All of these can be disabled via `SynthDef.newWithOptimizations`.

#### Sorting

Sorting uses a topological ordering that limits the number of wirebufs. This is done by the class `SynthDefTopologicalSort`.

#### CSE

CSE is done by the same class as subsitutions as they are related, that is the `SynthDefOptimizer`.

If a UGen with exactly the same edges and rate is found, and it is marked as `hasObservableEffect { ^false }` then it will be replaced.

This check is slow, particuarly for large graphs, but can offer great server--side performance boosts.

#### Subsitutions

Previously, UGens were able to alter any UGen in the graph, in any order.

Now, the optimiser works from the bottom of the graph to the top, stepping back down the graph when ever an optimisation is performed to check if anything else can be optimised. This can be slow if many optimisations take place, however this also provides large server--side speed up by reducing the number of UGens.

Although this isn't enforce, UGens are allowed to look at their **direct** inputs and children only to make decisions. Then the UGen must return a `SynthDefOptimisationResult` listing all the new UGens.

Importantly, UGens should **not** delete other UGens, instead, the dead-code elimination, which is performed automatically as a part of the topological sort, will clear these up. Except in the case where a UGen with `canBeReplacedByIdenticalCall { ^false }` needs to be deleted, in which case, it should be registered with the `SynthDefOptimisationResult` accordingly.

Additionally the `SynthDefOptimisationResult` should tell the optimiser what relative depth in the graph the changes occurred at, so that if the children were replaced, '1' is return, if the grand-children are replaced '2', etc..