plugins: TriggerUGens standardize init sample
adds internal state resets for the sake of enforcing conventions on
UGens initialization. This commit doesn't change any init sample or
output value.

When initializing a UGen, the convention is to reset internal state
after the first calc function call in Ctor. Even if such reset is not
absolutely necessary for these UGens, it makes it easier to reason about
initialization, by aligning with the convention.