reflect reviews and fixes bugs
1. Changes the way the rectangle is drawn to prevent collisions when there are many nodes. I initially tried to solve this problem by using a factor, but this approach was not very successful. I left this as a comment. The new implementation is not perfect, but it is an improvement on both the previous commit method and the abandoned approach.
2. It restricts multiple executions of the updater in a ServerTreeView. Starting again from the same ServerTreeView instance will be ignored and a warning will be displayed.
3. It adds a warning when multiple .plotTree calls are made.
4. Additional text is added to the current warning when the .plotTree function closes due to the server not running.

TO DO:
Allow the user to create multiple instances of ServerTreeView, including .plotTree. However, there is currently only one updater and one OSCFunc. Therefore, each instance cannot have its own interval, and stopping an instance does not update the label correctly. In fact, updating the node tree may occur when at least one ServerTreeView instance is running. This issue is documented in the example section of the ServerTreeView.schelp file. To resolve this issue, we need to discuss the various possible solutions.