Sclang: PyrGC: Insert limit on number of un-scanned objects.
When there are a large number of temporary objects mNumToScan can overflow leading to undefined behaviour (from the overflow of a signed int) and the implementation defined behaviour of right shifting a negative value, creating a GC bug that only appears on some systems, and incorrect, potentially leaking, but non-crashing behaviour on others.

This code cause a crash on some systems.
```supercollider
b = Pwhite.new(0, 10, inf).asStream;
Array.fill(1810000, {b.next});`
```

This commit makes mNumToScan a std::int64_t, which should be sufficient.

NOTE: if in future this bug reappears, place an upper limit on mNumToScan.