sclang: protect method lookup from segfault
sclang: protect method lookup from out of bounds index

When the method selector is a symbol that is a class name,
method lookup may be out of bounds. We check the symbol
and avoid the indexing.

This partly fixes #7007.

remove UNLIKELY, not necessary here.

sclang: protect doesNotUnderstand and post error if failed