classlib: DoesNotUnderstandError: refactor suggestions (#6997)
On the occasion of a refactoring of DoesNotUnderstandError this adds better introspection and guessing methods. They are added to Class and String so they can be used outside of DoesNotUnderstandError as well if needed.

We can now ask a class what messages its instances respond to and we can ask for similar strings in an array or similar selectors are in a class.

sclang: 
- add new primitive prInstancesOfClassRespondTo
- minor refactoring of variable scopes

new methods:
- String::findSimilarIn
- Class::respondingMethods
- Class::findSimilarSelectors
- Class::findRespondingSubclasses and findRespondingUpperSubclasses
- Class::instancesRespondTo (calls new primitive prInstancesOfClassRespondTo)

classlib: 
- refactor DoesNotUnderstandError
- DoesNotUnderstand error suggests all matching classes

minor fix:
- subclassResponsibility should pass method
- improve Class documantation

breaking change:
- DoesNotUnderstand suggestedCorrection getter removed