Improve and simplify musical example
The previous example could be naively solved by simply sorting the notes in each chord and assigning the lowest in source to lowest in target, next lowest to next lowest, etc.
This new example, although simpler, much better shows the power of solving via linear assignment because it demonstrates voice crossing, which would not be arrived at by the naive method.